+++ .gitignore
@@ -8,3 +8,4 @@ redisai.egg-info
 build/
 dist/
 docs/_build/
+.DS_Store
