+++ test/test.py
@@ -338,7 +338,8 @@ class DagTestCase(RedisAITestBase):
         self.assertEqual(expected, result)
         self.assertRaises(ResponseError, con.tensorget, 'b')
 
-    def test_dagrun_with_persist(self):
+    def dagrun_with_persist(self):
+        # TODO: disabling for now
         con = self.get_client()
 
         with self.assertRaises(ResponseError):
